// packages/core/src/@types/challengermode.d.ts
export interface GQLImage {
  /**
   * The url where the image is hosted.
   */
  url: string;

  /**
   * The image width.
   */
  width: number;

  /**
   * The image height.
   */
  height: number;
}
// ...
export interface GQLTeam {
  /**
   * The team logo image.
   */
  logo?: GQLImage;

  /**
   * The team banner image.
   */
  banner?: GQLImage;

  /**
   * The ID of the team.
   */
  id: GQLUUID;

  /**
   * The public name of the team.
   */
  name: string;

  /**
   * The public description of the team.
   */
  description: string;

  /**
   * Whether the team is verified.
   */
  verified: boolean;
}
// ...
export interface GQLTournamentLineup {
  /**
   * The team logo for this lineup. This will be available primarily if the lineup has a registered Team on challengermode.
   */
  logo?: GQLImage;

  /**
   * The team banner image.
   */
  banner?: GQLImage;

  /**
   * Get the persistent team that the lineup is playing as, if any.
   */
  team?: GQLTeam;

  /**
   * All members of this lineup.
   */
  members: Array<GQLTournamentLineupMember | null>;

  /**
   * The placement in the tournament for this lineup.
   *
   * If no results are available yet this will be null.
   */
  placement?: GQLTournamentLineupPlacement;

  /**
   * The participation status of the lineup.
   */
  participationStatus: GQLTournamentLineupParticipation;

  /**
   * The ID of the lineup.
   */
  id: GQLUUID;

  /**
   * The name of the tournament lineup. It may be sourced from a Challengermode team or party.
   * The value returned here is a snapshot, but it is not immutable - it might be adjusted when the name of a team changes globally, or it might not.
   */
  name: string;

  /**
   * Denotes if a lineup has been disqualified from the tournament. A disqualified lineup will concede all matches, and disappear from the bracket entirely.
   */
  disqualified: boolean;

  /**
   * A team's tournament seed is its ranking or placement in the tournament bracket, determined by automatic factors or manual input, to ensure fair and balanced matchups.This number is 0-indexed where lower means better. A team will always have if they are part of the rooster and tournament has started.
   */
  seed?: number;

  /**
   * A manual seed value that can be used during manual seeding to set the teams actual seed.
   */
  preSeed?: number;
}

/**
 * A member of a tournament lineup.
 */
export interface GQLTournamentLineupMember {
  /**
   * The in-game game account id of the member.
   *
   * Since this is game-title specific the format might vary.
   */
  gameAccountId: string;

  /**
   * A summary of member current status and available actions.
   */
  lobbyFeedback: GQLTournamentLobbyFeedback;

  /**
   * Whether the member is captain of the team.
   */
  captain: boolean;

  /**
   * The lineup role of the member
   * @deprecated No longer supported
   */
  role: string;

  /**
   * Whether the user has confirmed their participation.
   */
  confirmedTournamentParticipation: boolean;

  /**
   * The Challengermode profile of this user.
   */
  user: GQLUserProfile;
}

/**
 * A lineups current participation status in a tournament.
 */
export interface GQLTournamentLineupParticipation {
  /**
   * Whether the lineup has a ongoing match in the tournament.
   */
  matchRunning: boolean;

  /**
   * Whether the lineups participation in the tournament is complete.
   */
  participationComplete: boolean;

  /**
   * Whether the lineup is waiting for their next match in the tournament.
   */
  waitingForNextMatch: boolean;

  /**
   * The start time of the next match scheduled for the lineup in the tournament.
   */
  nextMatchScheduledAt?: GQLDateTime;
}

/**
 * Placement for a team in the tournament.
 *
 * Depending on the resolution of the tournament this is a range. E.g. If the tournament doesn't have bronze medal match, the semifinalist loosers will both have placement 3-4.
 */
export interface GQLTournamentLineupPlacement {
  /**
   * The best placement in the placement range.
   */
  bestPlacement: number;

  /**
   * The worst placement in the placement range.
   */
  worstPlacement: number;

  /**
   * A display text showing the placement range in which the team performed.
   */
  displayPlacement: string;
}
// ...
export interface GQLTournamentLobbyFeedback {
  /**
   * The current updated list of available actions for the user in the tournament.
   */
  allowedActions: Array<GQLAllowedTournamentAction>;

  /**
   * Translation key that can be used to describe the current state.
   */
  feedbackTranslationKey: string;

  /**
   * If the first action has a deadline, this time shows when it will expire.
   */
  timer?: GQLDateTime;
}
// ...
export interface GQLUserProfile {
  /**
   * Whether this is a BOT user.
   */
  bot: boolean;

  /**
   * When the user joined Challengermode.
   */
  createdAt: GQLDateTime;

  /**
   * The user's own description of themselves.
   */
  biography: string;

  /**
   * The user's username on Challengermode.
   */
  username: string;

  /**
   * The user's profile picture on Challengermode.
   */
  profilePicture?: GQLImage;

  /**
   * User statistics of their activity on Challengermode.
   */
  statistics?: GQLUserStatistics;

  /**
   * A list of the external accounts that the user has connected to their Challengermode account. Note that the user may decide to hide specific connected accounts.
   */
  connectedAccounts: Array<GQLConnectedAccount>;

  /**
   * List of all the game accounts that the user has added to their Challengermode accounts and is ready to participate in competitions in.
   */
  gameAccounts?: GQLGameAccountsConnection;

  /**
   * Whether or not the user is banned on Challengermode.
   */
  banned: boolean;

  /**
   * The user's Challengermode user ID.
   */
  userId: GQLUUID;

  /**
   * The user's username on Challengermode.
   */
  profileUrl: string;
}
// ...
export type GQLDateTime = any;
// ...
export type GQLUUID = any;